# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bid Management API",
        "description": "API для подачи и обработки заявок пользователями и администрацией.",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Основной сервер API"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/bids": {
            "get": {
                "tags": [
                    "Bids"
                ],
                "summary": "Получить список заявок",
                "operationId": "261fda4b83c5d8b30158533fc6b6c9ac",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_dir",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BidCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ"
                    },
                    "403": {
                        "description": "Доступ запрещён"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bids"
                ],
                "summary": "Создание новой заявки",
                "description": "Создание заявки доступно только авторизированным пользователям",
                "operationId": "852699b0f4d07d989ba550b5b905ad88",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "message"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Иван"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "ivan@example.com"
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "Мне нужна консультация"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Заявка успешно создана"
                    },
                    "403": {
                        "description": "Доступ запрещён"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/bids/{id}": {
            "get": {
                "tags": [
                    "Bids"
                ],
                "summary": "Получить заявку по ID",
                "operationId": "af13bce4c6b1ac2b10a0e946e764ca40",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ"
                    },
                    "404": {
                        "description": "Заявка не найдена"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Bids"
                ],
                "summary": "Обновление заявки",
                "description": "Обновление заявки доступно только админам и модераторам",
                "operationId": "1c11e534cbc64f6a19082addacc0bfcc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "Resolved, Reject"
                                    },
                                    "comment": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "Комментарий"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Заявка успешно обновлена"
                    },
                    "403": {
                        "description": "Доступ запрещён"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Bids"
                ],
                "summary": "Удаление заявки",
                "description": "Удаление заявки доступно только админам",
                "operationId": "cb1a4722040cfdc9ed9f1536b34d897e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Заявка успешно удалена"
                    },
                    "403": {
                        "description": "Доступ запрещён"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BidCollection": {
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/BidResource"
                }
            },
...
            "BidResource": {
                "required": [
                    "name",
                    "email",
                    "message",
                    "status",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string",
                        "example": "Иван"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "ivan@example.com"
                    },
                    "message": {
                        "type": "string",
                        "example": "Мне нужна консультация"
                    },
                    "status": {
                        "type": "string",
                        "example": "Active"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}